py-connect-postgress/requirements.txt
psycopg2-binary~=2.8.6